Rename tasks to task
